% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.TS.R
\name{get.TS}
\alias{get.TS}
\title{Reconstruct water level}
\usage{
get.TS(
  dat,
  init.h = 0,
  init.logsigmarw = 0,
  init.logSigma = getSigmaInit(dat$satid),
  bias = rep(0, length(unique(dat$satid)) - 1),
  init.logit = log(0.1/(1 - 0.1)),
  priorHeight = numeric(0),
  priorSd = numeric(0),
  estP = FALSE,
  silent = TRUE,
  weights = rep(1, nrow(dat)),
  varPerTrack = FALSE,
  varPerQuality = FALSE,
  newdat = NULL,
  ...
)
}
\arguments{
\item{dat}{A data.frame containing at least the columns: time, height, and track}

\item{init.h}{Initial value for the mean water levels. The default value is 0.}

\item{init.logsigmarw}{Initial value for the log of the standard deviation of the random walk}

\item{init.logSigma}{Initial value for the log of the standard deviation of the random walk}

\item{bias}{Optional, vector of length N-1 with Initial values of the bias estimates, where N is the numbers of satellite missions used. To estimate the bias "dat" must have a column "satid" with the ids of the satellites for each observation, 0,1,2,3,..,N-1. The number of bias estimates is N-1. The bias estimates is w.r.t. the satellite with the largest id. If dat$satid is provided the observation standard deviation is estimated pr satellite}

\item{init.logit}{Initial value for the log of the standard deviation of the observation noise}

\item{priorHeight}{...}

\item{priorSd}{...}

\item{estP}{A logic value \code{FALSE} if the outlier fraction is kept fixed at its initial value}

\item{weights}{Optional vector of weights.}

\item{varPerTrack}{Optional, a logic value: If TRUE, an observation standard deviation is "logSigma" is estimated per track.}

\item{varPerQuality}{Optional, a logic value: If TRUE, an observation standard deviation is "logSigma" is estimated per quality id. If this option is used "dat" must have a column named "qf"}

\item{newdat}{Optional, a data frame which at least should include a column named "time", containing the time in decimal years where the modeled water level is predicted. newdat may also include a column named "group" with a group id for each observation. Groups could be based on month, years, or something else. If "group" is provided a average water level pr group is also provided}

\item{exportPar}{Logic variable to specify if the estimated model parameters are saved to a file "tsPar.dat".}
}
\value{
An object of class "tsHydro" with the following elements
\itemize{
\item{oupfile("ts.dat") } A text file that contains three colunms;
"time", "wl", "wlsd". "time" is the time of each pass, where the water
level is estimated. "wl" is the estimated water level and "wlsd" is
the standard deviation of the estimated water level.
\item{"tspar.dat" } A text file that contains the optimized model parameters
}
}
\description{
Estimate the model parameters and return the estimated water levels
}
\details{
The function can handle the observation based standard deviation in different ways; either pr satellite, pr track, or pr quality. However, these options cannot be used together.
}
\examples{
data(lakelevels)
fit<-get.TS(lakelevels)


}
\keyword{series}
\keyword{time}
