\name{namco}
\alias{namco}
\docType{data}
\title{
namco
}
\description{
to be written
}
\usage{data("namco")}
\format{
  A data frame with 8476 observations on the following 4 variables.
  \describe{
    \item{\code{height}}{a numeric vector}
    \item{\code{lake}}{a numeric vector}
    \item{\code{track}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
  }
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(namco)
## maybe str(namco) ; plot(namco) ...
}
\keyword{datasets}
